server/pyproject.toml
[project]
name = "orchestra-api"
version = "0.1.0"
description = "REST API for managing RStudio workshops via Orchestra Operator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "kubernetes>=29.0.0",
    "python-dateutil>=2.8.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.1",
    "pyjwt[crypto]>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",  # For testing FastAPI
    "black>=24.0.0",
    "isort>=5.12.0",
    "ruff>=0.3.0",
]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py313"
